refactor : 인증 화면 로컬 컴포넌트 파일 분리
* 회원가입 퍼널 본체를 페이지에서 _components로 분리
* 자격 인증 화면과 심사 현황 화면의 콘텐츠 컴포넌트를 _components로 분리
* 헤더 브랜드 마크와 폼 카드를 형제 파일로 분리
* 소셜 로그인 콜백의 안내 레이아웃을 CenteredMessage로 분리
