+++ apps/web/src/app/(auth)/login/oauth2/code/[provider]/_components/CenteredMessage.tsx
@@ -1,0 +1,9 @@
+import type { ReactNode } from "react";
+
+export function CenteredMessage({ children }: { children: ReactNode }) {
+  return (
+    <div className="mx-auto flex min-h-dvh w-full max-w-md flex-col items-center justify-center gap-4 px-6 text-center">
+      {children}
+    </div>
+  );
+}
